requirements.txt
twitter==1.17.0
facebook-sdk
